+++ src/service/index.ts
@@ -22,3 +22,4 @@ export * from './SshpkService';
 export * from './VerifyService';
 export * from './VotingService';
 export * from './VotingUtils';
+export * from './ZipUtils';